backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
httpx==0.27.2
pydantic==2.9.2
sse-starlette==2.1.0
websockets==12.0
psutil==6.1.0
pyyaml==6.0.2
